refactor: split calendar grid into view components
