+++ src/components/calendar/TimeGrid.test.tsx
@@ -1,6 +1,6 @@
 // @vitest-environment jsdom
 import React from 'react';
-import { act, cleanup, render, screen, within } from '@testing-library/react';
+import { act, cleanup, render, screen } from '@testing-library/react';
 import { afterAll, afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import * as matchers from '@testing-library/jest-dom/matchers';
 expect.extend(matchers);
@@ -29,67 +29,10 @@ vi.mock('@dnd-kit/core', () => {
   };
 });
 
-
-import { CalendarGrid } from './CalendarGrid';
+import { TimeGrid } from './TimeGrid';
 import { DndContext } from '@dnd-kit/core';
 
-describe('CalendarGrid month view', () => {
-  afterEach(() => {
-    cleanup();
-    vi.useRealTimers();
-  });
-
-  it.each([
-    ['2023-02-01', 28],
-    ['2024-04-01', 30],
-    ['2024-05-01', 31],
-  ])('renders all days for %s', (iso, count) => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date(iso));
-    render(<CalendarGrid view="month" startOfWeek={new Date(iso)} events={[]} onDropTask={() => {}} />);
-    expect(screen.getAllByTestId('day-cell')).toHaveLength(count);
-  });
-
-  it('renders events inside the correct month day cell', () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date('2024-05-05T12:00:00Z'));
-    const events = [
-      { id: 'e1', taskId: 't1', title: 'Math HW', startAt: '2024-05-10T09:00:00.000Z', endAt: '2024-05-10T10:00:00.000Z' },
-      { id: 'e2', taskId: 't2', title: 'Science Project', startAt: '2024-05-10T14:00:00.000Z', endAt: '2024-05-10T15:00:00.000Z' },
-      { id: 'e3', taskId: 't3', title: 'Piano', startAt: '2024-05-12T09:00:00.000Z', endAt: '2024-05-12T10:00:00.000Z' },
-    ];
-    render(<CalendarGrid view="month" startOfWeek={new Date('2024-05-05T12:00:00Z')} events={events} onDropTask={() => {}} />);
-    expect(screen.getByText('Math HW')).toBeInTheDocument();
-    expect(screen.getByText('Science Project')).toBeInTheDocument();
-    expect(screen.getByText('Piano')).toBeInTheDocument();
-
-    const may10Cell = screen.getByLabelText('month-day-2024-05-10');
-    const may12Cell = screen.getByLabelText('month-day-2024-05-12');
-    expect(may10Cell.querySelectorAll('[data-testid="month-event"]').length).toBe(2);
-    expect(may12Cell.querySelectorAll('[data-testid="month-event"]').length).toBe(1);
-  });
-
-  it('renders multi-day events across days', () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date('2024-05-05T12:00:00Z'));
-    const events = [
-      {
-        id: 'm1',
-        taskId: 't1',
-        title: 'Retreat',
-        startAt: '2024-05-10T09:00:00.000Z',
-        endAt: '2024-05-12T10:00:00.000Z',
-      },
-    ];
-    render(<CalendarGrid view="month" startOfWeek={new Date('2024-05-05T12:00:00Z')} events={events} onDropTask={() => {}} />);
-    ['2024-05-10', '2024-05-11', '2024-05-12'].forEach((d) => {
-      const cell = screen.getByLabelText(`month-day-${d}`);
-      expect(within(cell).getByText('Retreat')).toBeInTheDocument();
-    });
-  });
-});
-
-describe('CalendarGrid interactions', () => {
+describe('TimeGrid interactions', () => {
   beforeEach(() => {
     vi.useFakeTimers();
     vi.setSystemTime(new Date('2099-01-01T00:00:00Z'));
@@ -113,7 +56,7 @@ describe('CalendarGrid interactions', () => {
           onDropTask(taskId, new Date(iso));
         }}
       >
-        <CalendarGrid view="day" startOfWeek={new Date('2099-01-01T00:00:00Z')} events={[]} onDropTask={onDropTask} />
+        <TimeGrid view="day" startOfWeek={new Date('2099-01-01T00:00:00Z')} events={[]} onDropTask={onDropTask} />
       </DndContext>
     );
     const iso = new Date('2099-01-01T08:00:00.000Z').toISOString();
@@ -126,7 +69,7 @@ describe('CalendarGrid interactions', () => {
     const start = '2099-01-01T09:00:00.000Z';
     const end = '2099-01-01T10:00:00.000Z';
     render(
-      <CalendarGrid
+      <TimeGrid
         view="day"
         startOfWeek={new Date('2099-01-01T00:00:00Z')}
         events={[{ id: 'e1', taskId: 't1', startAt: start, endAt: end, title: 'Study' }]}
@@ -142,28 +85,6 @@ describe('CalendarGrid interactions', () => {
     expect(onResizeEvent).toHaveBeenCalledWith('e1', 'end', new Date('2099-01-01T11:00:00.000Z'));
   });
 
-  it('updates event box position while dragging resize handles', () => {
-    const start = '2099-01-01T09:00:00.000Z';
-    const end = '2099-01-01T10:00:00.000Z';
-    render(
-      <CalendarGrid
-        view="day"
-        startOfWeek={new Date('2099-01-01T00:00:00Z')}
-        events={[{ id: 'e1', taskId: 't1', startAt: start, endAt: end, title: 'Study' }]}
-        onDropTask={() => {}}
-      />
-    );
-    const box = screen.getByText('Study').parentElement!.parentElement as HTMLElement;
-    expect(box.style.height).toBe('48px');
-    expect(box.style.top).toBe('432px');
-    act(() => {
-      const setter = transforms.get('event-resize-start-e1');
-      setter?.({ x: 0, y: -24 });
-    });
-    expect(box.style.top).toBe('408px');
-    expect(box.style.height).toBe('72px');
-  });
-
   it('stacks overlapping events in DOM order', () => {
     const events = [
       { id: 'a', taskId: 't1', startAt: '2099-01-01T09:00:00.000Z', endAt: '2099-01-01T10:00:00.000Z', title: 'Event A' },
@@ -171,7 +92,7 @@ describe('CalendarGrid interactions', () => {
     ];
     render(
       <DndContext>
-        <CalendarGrid view="day" startOfWeek={new Date('2099-01-01T00:00:00Z')} events={events} onDropTask={() => {}} />
+        <TimeGrid view="day" startOfWeek={new Date('2099-01-01T00:00:00Z')} events={events} onDropTask={() => {}} />
       </DndContext>
     );
     const aBox = screen.getByText('Event A');
@@ -182,7 +103,7 @@ describe('CalendarGrid interactions', () => {
   });
 });
 
-describe('CalendarGrid current time indicator', () => {
+describe('TimeGrid current time indicator', () => {
   afterEach(() => {
     cleanup();
     vi.useRealTimers();
@@ -191,20 +112,17 @@ describe('CalendarGrid current time indicator', () => {
   it('renders now line in day and week views', () => {
     vi.useFakeTimers();
     vi.setSystemTime(new Date('2099-01-01T09:00:00Z'));
-    render(<CalendarGrid view="day" startOfWeek={new Date('2099-01-01T00:00:00Z')} events={[]} onDropTask={() => {}} />);
+    render(<TimeGrid view="day" startOfWeek={new Date('2099-01-01T00:00:00Z')} events={[]} onDropTask={() => {}} />);
     expect(screen.getByTestId('now-indicator')).toBeInTheDocument();
     cleanup();
-    render(<CalendarGrid view="week" startOfWeek={new Date('2099-01-01T00:00:00Z')} events={[]} onDropTask={() => {}} />);
+    render(<TimeGrid view="week" startOfWeek={new Date('2099-01-01T00:00:00Z')} events={[]} onDropTask={() => {}} />);
     expect(screen.getByTestId('now-indicator')).toBeInTheDocument();
-    cleanup();
-    render(<CalendarGrid view="month" startOfWeek={new Date('2099-01-01T00:00:00Z')} events={[]} onDropTask={() => {}} />);
-    expect(screen.queryByTestId('now-indicator')).toBeNull();
   });
 
   it('updates position every minute', () => {
     vi.useFakeTimers();
     vi.setSystemTime(new Date('2099-01-01T09:30:00Z'));
-    render(<CalendarGrid view="day" events={[]} onDropTask={() => {}} />);
+    render(<TimeGrid view="day" events={[]} onDropTask={() => {}} />);
     const line = screen.getByTestId('now-indicator') as HTMLElement;
     const firstTop = parseFloat(line.style.top);
     act(() => {
@@ -216,7 +134,7 @@ describe('CalendarGrid current time indicator', () => {
   });
 });
 
-describe('CalendarGrid work hours + today highlight', () => {
+describe('TimeGrid work hours + today highlight', () => {
   afterEach(() => {
     cleanup();
     vi.useRealTimers();
@@ -226,7 +144,7 @@ describe('CalendarGrid work hours + today highlight', () => {
     vi.useFakeTimers();
     vi.setSystemTime(new Date('2099-01-01T09:00:00Z'));
     render(
-      <CalendarGrid
+      <TimeGrid
         view="day"
         startOfWeek={new Date('2099-01-01T00:00:00Z')}
         events={[]}
@@ -237,7 +155,6 @@ describe('CalendarGrid work hours + today highlight', () => {
     );
     expect(screen.getByText('12 AM')).toBeInTheDocument();
     expect(screen.getByText('11 PM')).toBeInTheDocument();
-    // One highlight overlay for the single day
     expect(screen.getAllByTestId('work-hours-highlight')).toHaveLength(1);
   });
 
@@ -245,7 +162,7 @@ describe('CalendarGrid work hours + today highlight', () => {
     vi.useFakeTimers();
     vi.setSystemTime(new Date('2024-05-15T09:00:00Z'));
     render(
-      <CalendarGrid
+      <TimeGrid
         view="week"
         startOfWeek={new Date('2024-05-13T00:00:00Z')}
         events={[]}
@@ -254,17 +171,7 @@ describe('CalendarGrid work hours + today highlight', () => {
         workEndHour={18}
       />
     );
-    // 7 work-hour overlays (one per day)
     expect(screen.getAllByTestId('work-hours-highlight')).toHaveLength(7);
-    // Today header is marked with title
-    expect(screen.getByTitle('Today')).toBeInTheDocument();
-  });
-
-  it('highlights today in month view', () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date('2024-05-15T09:00:00Z'));
-    render(<CalendarGrid view="month" startOfWeek={new Date('2024-05-01T00:00:00Z')} events={[]} onDropTask={() => {}} />);
-    // Month view shows one element marked as Today
     expect(screen.getByTitle('Today')).toBeInTheDocument();
   });
 });
